+++ tests/integration_tests/tests_with_correlation/test_1D_1D_correlation_multiplicative_model_error.py
@@ -157,7 +157,7 @@ class TestProblem(unittest.TestCase):
             prior=Normal(mean=mean_EI, std=std_EI),
         )
         problem.add_parameter(
-            "L", "model", tex="$L$", info="Length of the beam [m]", const=L
+            "L", "model", tex="$L$", info="Length of the beam [m]", value=L
         )
         problem.add_parameter(
             name="sigma",