+++ tests/integration_tests/tests_with_correlation/test_1D_2D_correlation_additive_model_error.py
@@ -163,7 +163,7 @@ class TestProblem(unittest.TestCase):
             prior=Normal(mean=mean_EI, std=std_EI),
         )
         problem.add_parameter(
-            name="L", tex="$L$", info="Length of the beam [m]", const=L
+            name="L", tex="$L$", info="Length of the beam [m]", value=L
         )
         problem.add_parameter(
             name="sigma",